Changed Binary Search Tree, Red Black Tree tree nodes to be private inner classes instead of being declared in global namespace

+++ src/main/cpp/graphtheory/shortestpaths/BellmanFord.h
@@ -14,7 +14,7 @@ private:
   };
   
   // Each index represents predecessor and shortest path length to that vertex
-  typedef std::vector<std::pair<int, int>> returnType;
+  typedef std::vector<std::pair<int, double>> returnType;
   typedef std::vector<std::list<std::pair<int, double>>> inputAdjListType;
   typedef std::vector<std::list<std::pair<vertex, double>>> adjListType;
   typedef std::vector<vertex> vertListType;
